+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.2.0
 
 - Exit with code 1 on error, including reference errors
+- Support source map references
 
 # 0.1.1